+++ src/careamics_restoration/config/algorithm.py
