+++ src/careamics_restoration/config/training.py
