+++ src/careamics_restoration/manipulation/manipulation_factory.py
