+++ src/careamics_restoration/manipulation/pixel_manipulation.py
