+++ src/careamics_restoration/models/model_factory.py
