+++ src/careamics_restoration/models/models.py
