--- src/n2v/training.py
@@ -1,1 +1,0 @@
-#TODO train single epoch should be here
