+++ tests/test_metrics.py
@@ -2,7 +2,7 @@ import pytest
 
 import numpy as np
 
-from n2v.metrics import (
+from careamics_restoration.metrics import (
     zero_mean,
     scale_invariant_psnr,
     MetricTracker,
@@ -10,22 +10,28 @@ from n2v.metrics import (
 
 # TODO tests fix and fix_range
 
-@pytest.mark.parametrize("x", [
+
+@pytest.mark.parametrize(
+    "x",
+    [
         5.6,
         np.array([1, 2, 3, 4, 5]),
         np.array([[1, 2, 3], [4, 5, 6]]),
-    ]
+    ],
 )
 def test_zero_mean(x):
-    assert np.allclose(zero_mean(x), x-np.mean(x))
+    assert np.allclose(zero_mean(x), x - np.mean(x))
 
 
-@pytest.mark.parametrize("gt, pred, result", [
+@pytest.mark.parametrize(
+    "gt, pred, result",
+    [
         (np.array([1, 2, 3, 4, 5, 6]), np.array([1, 2, 3, 4, 5, 6]), 332.22),
         (np.array([[1, 2, 3], [4, 5, 6]]), np.array([[1, 2, 3], [4, 5, 6]]), 332.22),
-])
+    ],
+)
 def test_scale_invariant_psnr(gt, pred, result):
-    assert scale_invariant_psnr(gt, pred) == pytest.approx(result, rel = 5e-3)
+    assert scale_invariant_psnr(gt, pred) == pytest.approx(result, rel=5e-3)
 
 
 def test_metric_tracker():
@@ -43,7 +49,7 @@ def test_metric_tracker():
         tracker.update(i, n)
 
     # check values
-    assert tracker.sum == n * ( n * (n-1)) / 2
-    assert tracker.count == n*n
-    assert tracker.avg == (n-1) / 2
-    assert tracker.val == n-1
+    assert tracker.sum == n * (n * (n - 1)) / 2
+    assert tracker.count == n * n
+    assert tracker.avg == (n - 1) / 2
+    assert tracker.val == n - 1
